--- a/uproot_methods/version.py
+++ b/uproot_methods/version.py
@@ -30,7 +30,7 @@
 
 import re
 
-__version__ = "0.2.9"
+__version__ = "0.2.10"
 version = __version__
 version_info = tuple(re.split(r"[-\.]", __version__))