A small improvement in eta evaluation using `asinh` (`atanh` would save a `pt` call but it would be numerically imprecise at large `z` due to the asymptote)
```python
entries = 70000
p3 = np.random.uniform(0, 20, size=entries*3).reshape(-1,3)
E = np.sqrt(np.sum(p3**2, axis=1))
p4 = uproot_methods.TLorentzVectorArray(p3[:,0], p3[:,1], p3[:,2], E)

%timeit p4.eta
```
gives
```
Old: 5.79 ms ± 223 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
New: 2.53 ms ± 114 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
I also made a `memoized` decorator class, but I'm not sure if this is the correct idiom.  Consider it WIP for now