@jpivarski 2018-12-02
Yeah, this is gangbusters! Just the mathematical simplification alone is a big deal (isolating it out) is a factor of 22. Full disclosure: I haven't done any performance tests like this in the development of this library; just made sure they were all Numpy calls. Thanks for finding this!

(Note: I'm also avoiding the creation of an intermediate `TVector3` with `.p3`, too.)

```
% ipython -i -c 'import numpy; x = numpy.random.normal(0, 1, 100000); y = numpy.random.normal(0, 1, 100000); z = numpy.random.normal(0, 1, 100000);'
Python 3.7.0 | packaged by conda-forge | (default, Sep 30 2018, 14:56:18)
Type 'copyright', 'credits' or 'license' for more information
IPython 6.5.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: %%timeit
   ...: -numpy.log((1.0 - numpy.cos(numpy.arctan2(numpy.sqrt(x**2 + y**2), z))) / (1.0 + numpy.
   ...: cos(numpy.arctan2(numpy.sqrt(x**2 + y**2), z)))) / 2.0
   ...:
8.13 ms ± 702 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [2]: %%timeit
   ...: -numpy.log((1.0 - numpy.cos(numpy.arctan2(numpy.sqrt(x**2 + y**2), z))) / (1.0 + numpy.
   ...: cos(numpy.arctan2(numpy.sqrt(x**2 + y**2), z)))) / 2.0
   ...:
8.53 ms ± 368 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [3]: %%timeit
   ...: numpy.arcsinh(z / numpy.sqrt(x**2 + y**2))
   ...:
357 µs ± 42.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [4]: %%timeit
   ...: numpy.arcsinh(z / numpy.sqrt(x**2 + y**2))
   ...:
365 µs ± 25.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [5]: 8.2 / 0.36
Out[5]: 22.777777777777775

In [6]: ((-numpy.log((1.0 - numpy.cos(numpy.arctan2(numpy.sqrt(x**2 + y**2), z))) / (1.0 + nump
   ...: y.cos(numpy.arctan2(numpy.sqrt(x**2 + y**2), z)))) / 2.0) - (numpy.arcsinh(z / numpy.sq
   ...: rt(x**2 + y**2))) > 1e-11).any()
Out[6]: False
```

I like the idea of memoization, too, and we can expand its use so that constructors from cylindrical coordinates preload the memo. Another user has already complained that when they construct an array with a known set of masses, the masses from `.mass` differ in the last decimal point, because we are, in fact, recomputing it. I'd also like to put the memoization at a higher level, so that it's a general facility.